% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empRank.R
\name{empRank}
\alias{empRank}
\title{Compute empirical ranks}
\usage{
empRank(Ly, Lt, workGrid = NULL)
}
\arguments{
\item{Ly}{A list of \eqn{n} vectors holding the observed values for each subject.}

\item{Lt}{A list of \eqn{n} vectors holding the observation time points for each subject. 
Each vector should be in ascending order and of the same length as the corresponding element in \code{Ly}.}

\item{workGrid}{A vector holding the time grid where the empirical ranks are computed. Default: set up by \code{\link{setWorkGrid}}.}
}
\value{
An \eqn{n}-by-\eqn{l} matrix of which each column holds the empirical ranks for one subject evaluated on \code{workGrid},
where \eqn{l} is the length of \code{workGrid}.
}
\description{
Compute empirical ranks
}
