% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerfctn.R
\name{kerfctn}
\alias{kerfctn}
\title{Kernel function}
\usage{
kerfctn(kernel_type = "epan", fctn_type = "pdf")
}
\arguments{
\item{kernel_type}{Character holding the kernel type: \code{"rect"}, \code{"gaus"}, \code{"epan"} (default), \code{"tgaus"}.}

\item{fctn_type}{Character holding the function type: \code{"pdf"} (default) , \code{"cdf"}, \code{"dpdf"} (derivatives of pdfs).}
}
\value{
A function computing the kernel function.
}
\description{
Kernel function
}
