% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeGrid.R
\name{normalizeGrid}
\alias{normalizeGrid}
\title{Normalize time grid}
\usage{
normalizeGrid(tgrid)
}
\arguments{
\item{tgrid}{A vector holding the time grid to be normalized.}
}
\value{
A grid on \eqn{[0,1]}.
}
\description{
Normalize time grid
}
