% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankDerivSql2.R
\name{rankDerivSql2}
\alias{rankDerivSql2}
\title{Compute subject-specific squared \eqn{L^2} norm of the rank derivatives}
\usage{
rankDerivSql2(rankDeriv, tgrid)
}
\arguments{
\item{rankDeriv}{A vector of length \eqn{l} or an \eqn{n}-by-\eqn{l} matrix holding the rank derivatives evaluated on \code{tgrid}.}

\item{tgrid}{A vector of length \eqn{l} holding the time grid where the rank derivatives are evaluated.}
}
\value{
A vector holding the squared \eqn{L^2} norm of the rank derivatives for each subject.
}
\description{
Compute subject-specific squared \eqn{L^2} norm of the rank derivatives
}
