% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankPtDeriv.R
\name{rankPtDeriv}
\alias{rankPtDeriv}
\title{Compute rank partial derivatives}
\usage{
rankPtDeriv(Ly, Lt, optns = list())
}
\arguments{
\item{Ly}{A list of \eqn{n} vectors holding the observed values for each subject.}

\item{Lt}{A list of \eqn{n} vectors holding the observation time points for each subject. 
Each vector should be in ascending order and of the same length as the corresponding element in \code{Ly}.}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}. 
Available control options are \code{tWorkGrid}, \code{yWorkMat}, \code{tbw}, \code{ybw}, 
\code{Kty}, \code{Hty}, \code{ncores}; see \code{\link{smoRank}} for details.}
}
\value{
A list of the following elements:
\describe{
\item{D1, D2}{Matrices of dimension \eqn{n}-by-\eqn{l} of which the \eqn{(i,j)}-th entry holds the corresponding derivatives for the \eqn{i}-th subject evaluated at \code{tWorkGrid[j]} and \code{yWorkMat[i,j]}.}
\item{optns}{A list of control options used in computing \code{D1} and \code{D2}.}
}
}
\description{
Compute rank partial derivatives \eqn{D_1(Y_i(t_j),t_j)} and \eqn{D_2(Y_i(t_j),t_j)}, 
where \eqn{D_1(y,t) = \partial F_t(y)/\partial t} and \eqn{D_2(y,t) = \partial F_t(y)/\partial y}.
}
