% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankStab.R
\name{rankStab}
\alias{rankStab}
\title{Compute overall population rank stability}
\usage{
rankStab(rankDeriv, tgrid)
}
\arguments{
\item{rankDeriv}{An \eqn{n}-by-\eqn{l} matrix holding the rank derivatives evaluated on \code{tgrid}.}

\item{tgrid}{A vector of length \eqn{l} holding the time grid where the rank derivatives are evaluated.}
}
\value{
A scalar holding the overall population rank stability.
}
\description{
Compute overall population rank stability
}
