% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setWorkGrid.R
\name{setWorkGrid}
\alias{setWorkGrid}
\title{Set working grid for a given list of time grids.}
\usage{
setWorkGrid(Lt, nGrid = 101)
}
\arguments{
\item{Lt}{A list of vectors holding the observed time points.}

\item{nGrid}{The number of equidistant grid points of the output grid. Default: 101.}
}
\value{
A vector holding an equidistant time grid.
}
\description{
Set working grid for a given list of time grids.
}
