% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoRank.R
\name{smoRank}
\alias{smoRank}
\title{Compute smooth ranks}
\usage{
smoRank(Ly, Lt, optns = list())
}
\arguments{
\item{Ly}{A list of \eqn{n} vectors holding the observed values for each subject.}

\item{Lt}{A list of \eqn{n} vectors holding the observation time points for each subject. 
Each vector should be in ascending order and of the same length as the corresponding element in \code{Ly}.}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}. See `Details'.}
}
\value{
A list of the following elements if \code{bwchoice} is \code{FALSE} and a list including \code{ybw} and \code{tbw} otherwise.
\describe{
\item{rank}{An \eqn{n}-by-\eqn{l} matrix of which the \eqn{(i,j)}-th entry holds the smooth ranks of the \eqn{i}-th subject evaluated at \code{tWorkGrid[j]} and \code{yWorkMat[i,j]}.}
\item{optns}{A list of control options used in computing \code{rank}.}
}
}
\description{
Compute smooth ranks
}
\details{
Available control options are 
\describe{
\item{tWorkGrid}{A vector of length \eqn{l} holding the time grid where the smooth ranks are computed.}
\item{yWorkMat}{An \eqn{n}-by-\eqn{l} matrix; each row holding the values of the process for one subject observed on \code{tWorkGrid} where the smooth ranks are computed.}
\item{tbw, ybw}{Positive scalars holding the bandwidths.}
\item{Kty, Hty}{A character string holding the type of kernel functions; \code{"rect"}, \code{"gaus"}, \code{"epan"}, \code{"tgaus"}; default: \code{"epan"}.}
\item{training}{A scalar between 0 and 1 holding the proportion of the subjects used for choosing the bandwidths; default: 1.}
\item{ncores}{Number of cores used in parallel; default: 1.}
\item{bwchoice}{logical. If \code{TRUE}, only return the choices for \code{ybw} and \code{tbw}. Default: \code{FALSE}.}
}
}
