% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvRankStab.R
\name{tvRankStab}
\alias{tvRankStab}
\title{Compute time-dependent population rank stability}
\usage{
tvRankStab(rankDeriv)
}
\arguments{
\item{rankDeriv}{An \eqn{n}-by-\eqn{l} matrix holding the rank derivatives evaluated on \code{tgrid}.}
}
\value{
A vector of length \eqn{l} holding the time-dependent population rank stability.
}
\description{
Compute time-dependent population rank stability
}
